+++ kernel/devices/kybrd.c
@@ -646,7 +646,7 @@ static int kybrd_open(struct vfs_inode *inode, struct vfs_file *file) {
 	return 0;
 }
 
-static ssize_t kybrd_read(struct vfs_file *file, char *buf, size_t count, off_t ppos) {
+static uint32_t kybrd_read(struct vfs_file *file, char *buf, size_t count, off_t ppos) {
 	struct kybrd_inode *mi = (struct kybrd_inode *)file->private_data;
 	wait_event(&hwait, mi->ready);
