+++ libc/include/sys/types.h
@@ -9,7 +9,7 @@ typedef long long off_t;
 typedef long off_t;
 typedef unsigned long long ino_t;
 typedef unsigned int mode_t;
-typedef long ssize_t;
+typedef long uint32_t;
 typedef unsigned long sector_t;
 typedef int pid_t;
 typedef int tid_t;
